fix: renamed address component
